third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/tables/hidden-attr-expected.txt

This is a testharness.js-based test.
PASS Computed display and visibility of table
PASS Computed display and visibility of caption
FAIL Computed display and visibility of colgroup assert_equals: expected "table-column-group" but got "none"
FAIL Computed display and visibility of col assert_equals: expected "table-column" but got "none"
FAIL Computed display and visibility of thead assert_equals: expected "table-header-group" but got "none"
FAIL Computed display and visibility of tbody assert_equals: expected "table-row-group" but got "none"
FAIL Computed display and visibility of tfoot assert_equals: expected "table-footer-group" but got "none"
FAIL Computed display and visibility of tr assert_equals: expected "table-row" but got "none"
PASS Computed display and visibility of td
PASS Computed display and visibility of th
Harness: the test ran to completion.
